chore(*) generate the remaining mesh resources
Generate the remaining mesh resource wrappers that were overlooked in
the previous pass.
