+++ api/mesh/v1alpha1/health_check.proto
@@ -4,8 +4,8 @@ package kuma.mesh.v1alpha1;
 
 option go_package = "github.com/kumahq/kuma/api/mesh/v1alpha1";
 
+import "mesh/options.proto";
 import "mesh/v1alpha1/selector.proto";
-
 import "google/protobuf/duration.proto";
 import "google/protobuf/wrappers.proto";
 import "validate/validate.proto";
@@ -19,6 +19,11 @@ option (doc.config) = {
 
 // HealthCheck defines configuration for health checking.
 message HealthCheck {
+
+  option (kuma.mesh.resource).name = "HealthCheckResource";
+  option (kuma.mesh.resource).type = "HealthCheck";
+  option (kuma.mesh.resource).package = "mesh";
+
   // List of selectors to match dataplanes that should be configured to do
   // health checks.
   repeated Selector sources = 1 [ (validate.rules).repeated .min_items = 1 ];
