+++ api/mesh/v1alpha1/zone_ingress_insight.proto
@@ -4,10 +4,17 @@ package kuma.mesh.v1alpha1;
 
 option go_package = "github.com/kumahq/kuma/api/mesh/v1alpha1";
 
+import "mesh/options.proto";
 import "mesh/v1alpha1/dataplane_insight.proto";
 
 // ZoneIngressInsight defines the observed state of a Zone Ingress.
 message ZoneIngressInsight {
+
+  option (kuma.mesh.resource).name = "ZoneIngressInsightResource";
+  option (kuma.mesh.resource).type = "ZoneIngressInsight";
+  option (kuma.mesh.resource).package = "mesh";
+  option (kuma.mesh.resource).global = true;
+  option (kuma.mesh.resource).skip_validation = true;
 
   // List of ADS subscriptions created by a given Zone Kuma CP.
   repeated DiscoverySubscription subscriptions = 1;
